+++ c-server/_utils.h
@@ -2,6 +2,6 @@
 #define _UTILS_H
 
 size_t build_response(char ** response);
-size_t extract_header(char * reqContent, char * headerName, char ** output);
+int extract_header(char * reqContent, char * headerName, char ** output);
 
 #endif
